+++ docs/source/changelog.rst
@@ -6,6 +6,7 @@ Paver Changelog
 1.0.4 (unreleased)
 ------------------
 * Fixed md5 deprecation warnings in the bundled cog (thanks to jszakmeister, issue #56)
+* Tasks take cmdopts even from grandparents (thanks to aurelianito, github issue #4)
 
 1.0.3 (June 1, 2010)
 --------------------